requirements.txt
anyio==3.6.1
attrs==21.4.0
autobahn==22.6.1
Automat==20.2.0
binance-connector==1.14.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
click==8.1.3
client==0.0.0
colorama==0.4.5
constantly==15.1.0
cryptography==37.0.4
dnspython==2.2.1
email-validator==1.2.1
fastapi==0.79.0
greenlet==1.1.2
h11==0.13.0
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
passlib==1.7.4
psycopg2-binary==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydantic==1.9.1
PyJWT==2.4.0
pyOpenSSL==22.0.0
python-decouple==3.6
python-dotenv==0.20.0
python-multipart==0.0.5
requests==2.28.1
service-identity==21.1.0
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.39
sqlalchemy-json==0.5.0
starlette==0.19.1
Twisted==22.4.0
txaio==22.2.1
typing_extensions==4.3.0
urllib3==1.26.10
uvicorn==0.18.2
zope.interface==5.4.0
